feat: unify disciple badge ui
