+++ docs/00-wireframes.md
@@ -80,7 +80,7 @@ Water Orb widget showing current Water per second.
 Footer
 
 “Gate” button to open the full map/exploration/tasks overlay.
-* Exploration disciple selection now uses bamboo-colored badges with a small mood icon.
+* Exploration and sect disciple lists now use a unified bamboo-colored badge showing the disciple's name, mood, HP bar and current task.
 
 #dicsiple overlay card on pressing disciple
